Empty message
git-svn-id: https://svn.lvk.cs.msu.su/axiomlib/branches/armkor@152 08e97c73-d1ae-4023-87e5-8c5be0462271

+++ src/RecognizerMultimarkup.h
@@ -25,7 +25,7 @@ private:
 
 
 
-    int stretch; // используется при указании размеров окна
+    double stretch; // используется при указании размеров окна
     MultiMarking::DTWMetric* metric;//используемая метрика
     std::string name_metric; // name of metric
